Breaking up the utils.tx into multiple smaller source files and adding a filemod check function to prevent excessive compiling on contracts
