+++ src/cli/cli-utils/logIndicator.ts
