+++ src/index.ts
@@ -3,5 +3,5 @@ export * from './contracts';
 export * from './eosManager';
 export * from './utils';
 
-import * as cliUtils from './cli/utils';
+import * as cliUtils from './cli/cli-utils/cli-utils';
 export const CLI = cliUtils;
